[AIRFLOW-5937]  Group tests for the version/connection/db commands
